requirements.txt
# Construction PM - Backend dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.9
pillow==10.2.0
weasyprint==60.2
reportlab>=4.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.4
